Extract email utils into a separate file
